% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alluvialplot}
\alias{alluvialplot}
\title{Create alluvialplot.}
\usage{
alluvialplot(
  .data,
  .x,
  .label = NULL,
  .fill = NULL,
  .title = NULL,
  .gridrow = NULL,
  .gridcol = NULL,
  .fillcolors = NULL
)
}
\arguments{
\item{.data}{tibble or data_frame with data}

\item{.x}{column defining x-axis}

\item{.label}{optional x-axis label}

\item{.fill}{optional column for subgroups}

\item{.title}{optional plot title}

\item{.gridrow}{optional facetting variable}

\item{.gridcol}{optional facetting variable}

\item{.fillcolors}{values manually defining fill colors.}
}
\description{
\code{alluvialplot} returns a ggplot object.
}
