% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{cat_desc_stats}
\alias{cat_desc_stats}
\title{Compute absolute and relative frequencies.}
\usage{
cat_desc_stats(
  quelle,
  separator = " ",
  return_level = TRUE,
  ndigit = 0,
  groupvar = NULL,
  singleline = FALSE,
  percent = TRUE,
  prettynum = FALSE,
  .german = FALSE
)
}
\arguments{
\item{quelle}{Data for computation.}

\item{separator}{delimiter between results per level, preset as ' '.}

\item{return_level}{Should levels be reported?}

\item{ndigit}{Digits for rounding of relative frequencies.}

\item{groupvar}{Optional grouping factor.}

\item{singleline}{Put all group levels in  a single line?}

\item{percent}{Logical, add percent-symbol after relative frequencies?}

\item{prettynum}{logical, apply prettyNum to results?}

\item{.german}{logical, should "." and "," be used as bigmark and decimal? Sets prettynum to TRUE}
}
\description{
Compute absolute and relative frequencies.
}
