% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{cn}
\alias{cn}
\title{Shortcut for colnames()}
\usage{
cn(data = rawdata)
}
\arguments{
\item{data}{Data structure to read colnames from.}
}
\description{
\code{cn} lists colnames, by default for variable rawdata.
}
