% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{ColSeeker}
\alias{ColSeeker}
\title{Find numeric index and names of columns based on type and patterns}
\usage{
ColSeeker(
  data = rawdata,
  namepattern = ".",
  varclass = NULL,
  exclude = NULL,
  casesensitive = TRUE
)
}
\arguments{
\item{data}{tibble or data.frame, where columns are to be found; by default rawdata}

\item{namepattern}{Vector of pattern to look for.}

\item{varclass}{Only columns of defined type are returned}

\item{exclude}{Vector of pattern to exclude from found names.}

\item{casesensitive}{Logical if case is respected in matching (default FALSE: a<>A)}
}
\value{
A list with index, names, and backticked names
}
\description{
\code{ColSeeker} looks up colnames (by default for data-frame rawdata)
based on type and parts of names, using regular expressions.
Be warned that special characters as e.g. \code{[} \code{(} need to be escaped or replaced by \code{.}
Exclusion rules may be specified as well.
}
