% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare2qualvars}
\alias{compare2qualvars}
\title{Comparison for columns of factors for 2 groups}
\usage{
compare2qualvars(data, testvars, groupvar, round_p = 3, round_desc = 2,
  pretext = F, mark = F, singleline = F, newline = T)
}
\description{
Comparison for columns of factors for 2 groups
}
