% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare_n_numvars}
\alias{compare_n_numvars}
\title{Comparison for columns of Gaussian or ordinal measures for n groups}
\usage{
compare_n_numvars(
  .data = rawdata,
  dep_vars,
  indep_var,
  gaussian,
  round_desc = 2,
  range = FALSE,
  rangesep = " ",
  pretext = FALSE,
  mark = FALSE,
  round_p = 3,
  add_n = FALSE,
  ci = FALSE
)
}
\arguments{
\item{.data}{name of dataset (tibble/data.frame) to analyze, defaults to rawdata.}

\item{dep_vars}{vector of column names.}

\item{indep_var}{name of grouping variable.}

\item{gaussian}{Logical specifying normal or ordinal indep_var (and chooses comparison tests accordingly)}

\item{round_desc}{number of significant digits for rounding of descriptive stats.}

\item{range}{include min/max?}

\item{rangesep}{text between statistics and range or other elements.}

\item{pretext}{argument for function formatP, adding = or <.}

\item{mark}{argument for formatP, adding significance indicators.}

\item{round_p}{level for rounding p-value.}

\item{add_n}{add n to descriptive statistics?}

\item{ci}{include bootstrap CI for descriptive statistics? Defaults to FALSE}
}
\value{
A list with elements "results": tibble with descriptive statistics,
p-value from ANOVA/Kruskal-Wallis test, p-values for pairwise comparisons, significance
indicators, and descriptives pasted with significance.
Columns ending with \code{ fn} report descriptive statistics with post-hoc results as footnote.
"raw": nested list with output from all underlying analyses.
}
\description{
Some names were changed in August 2022, to reflect the update of the function to handle ordinal data using non-parametric equivalents.
}
\examples{
# Usually, only the result table is relevant:
compare_n_numvars(
  .data = mtcars, dep_vars = c("wt", "mpg", "hp"),
  indep_var = "cyl",
  gaussian = TRUE
)$results
# For a report, result columns may be filtered as needed:
compare_n_numvars(
  .data = mtcars, dep_vars = c("wt", "mpg", "hp"),
  indep_var = "cyl",
  gaussian = FALSE
)$results |>
  dplyr::select(Variable, `cyl 4 fn`:`cyl 8 fn`, multivar_p)
}
