% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare_n_qualvars}
\alias{compare_n_qualvars}
\title{Comparison for columns of factors for more than 2 groups with post-hoc}
\usage{
compare_n_qualvars(
  data,
  testvars,
  groupvar,
  round_p = 3,
  round_desc = 2,
  pretext = F,
  mark = F,
  singleline = F,
  spacer = "&nbsp;",
  linebreak = "ARRR",
  prettynum = F
)
}
\description{
Comparison for columns of factors for more than 2 groups with post-hoc
}
