% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{cortestR}
\alias{cortestR}
\title{correlation matrix with p-values based on cor.test}
\usage{
cortestR(cordata, method = "pearson", digits = 3, digits_p = 3,
  sign = "symbol")
}
\arguments{
\item{cordata}{data frame or matrix with rawdata.}

\item{method}{as in cor.test.}

\item{digits}{rounding level for estimate.}

\item{digits_p}{rounding level for p value.}

\item{sign}{if 'symbol', add significance indicator.}
}
\description{
\code{cortestR} returns a data frame with coefficient,p-values,
and significance symbols.
}
