% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Find outliers based on IQR}
\usage{
detect_outliers(x, coef = 1.5)
}
\arguments{
\item{x}{numeric vector.}

\item{coef}{coefficient for boxplot.stats, defaults to 1.5.}
}
\value{
A list with elements positions and outliers as numeric vectors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{detect_outliers} computes IQR and finds outliers. It gives the same results as \code{geom_boxplot} and thus differs slightly from \code{boxplot.stats}.
}
\examples{
detect_outliers(rnorm(100))
}
