% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{FindVars}
\alias{FindVars}
\title{Find numeric index and names of column name patterns}
\usage{
FindVars(varnames, allnames = colnames(rawdata), exact = F,
  exclude = NA, casesensitive = T, fixed = F)
}
\arguments{
\item{varnames}{Vector of pattern to look for.}

\item{allnames}{Vector of values to detect pattern in;
by default, colnames(rawdata).}

\item{exact}{Partial matching or exact only (adding ^ and $)?}

\item{exclude}{Vector of pattern to exclude from found names.}

\item{casesensitive}{Logical if case is respected in matching (default FALSE: a<>A)}

\item{regex}{Logical, treat pattern as regex?}
}
\value{
A list with index, names, backticked names, and symbols
}
\description{
\code{FindVars} looks up columnames (by default for data-frame rawdata)
based on parts of names, using regular expressions.
exlusion rules may be specified as well.
}
\examples{
FindVars(varnames = c('^c','g'),allnames = colnames(mtcars))
FindVars(varnames = c('^c','g'),allnames = colnames(mtcars),exclude='r')
}
