% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_rtree}
\alias{gg_rtree}
\title{Plot  a regression tree with ggplot.}
\usage{
gg_rtree(rpartdata, miny = NULL, title = "", german = FALSE)
}
\arguments{
\item{rpartdata}{output from rpart.}

\item{miny}{??.}

\item{title}{Headline for plot.}

\item{german}{Use German numbers.}
}
\description{
Plot  a regression tree with ggplot.
}
