% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ksnormal}
\alias{ksnormal}
\title{Convinience function around ks.test, testing against Normal distribution}
\usage{
ksnormal(ksdata)
}
\arguments{
\item{ksdata}{Vector of data to test.}
}
\description{
\code{ksnormal} returns list output from ks.test.
<<<<<<< HEAD
}
=======
}
>>>>>>> 4dafd7a42e9093a1c3f1d078eaca5e29a7d66907
