% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{logrange_123456789}
\alias{logrange_123456789}
\title{vector for log-scales at 1 to 9}
\format{
An object of class \code{numeric} of length 369.
}
\usage{
logrange_123456789
}
\description{
vector for log-scales at 1 to 9
}
\keyword{datasets}
