% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{logrange_12357}
\alias{logrange_12357}
\title{vector for log-scales at 1 ,2,3,5,7}
\format{
An object of class \code{numeric} of length 205.
}
\usage{
logrange_12357
}
\description{
vector for log-scales at 1 ,2,3,5,7
}
\keyword{datasets}
