% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{logrange_15}
\alias{logrange_15}
\title{vector for log-scales at 1 and 5}
\format{
An object of class \code{numeric} of length 82.
}
\usage{
logrange_15
}
\description{
vector for log-scales at 1 and 5
}
\keyword{datasets}
