% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{logrange_5}
\alias{logrange_5}
\title{vector for log-scales with step .5}
\format{
An object of class \code{numeric} of length 738.
}
\usage{
logrange_5
}
\description{
vector for log-scales with step .5
}
\keyword{datasets}
