% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{markSign}
\alias{markSign}
\title{Convert significance levels to symbols}
\usage{
markSign(SignIn, plabel = c("n.s.", "+", "*", "**", "***"))
}
\arguments{
\item{SignIn}{A single p-value.}

\item{plabel}{A translation table, predefined with the usual symbols.}
}
\value{
factor with label as defined in plabel.
}
\description{
\code{markSign} returns the symbol associated with a significance level.
}
\examples{
markSign(c(0.12, 0.012, 0.0012))
}
