% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{meansd}
\alias{meansd}
\title{Compute mean and sd and put together with the +- symbol.}
\usage{
meansd(
  x,
  roundDig = 2,
  drop0 = FALSE,
  groupvar = NULL,
  range = FALSE,
  rangesep = " ",
  .n = FALSE,
  .german = FALSE
)
}
\arguments{
\item{x}{Data for computation.}

\item{roundDig}{Number of relevant digits for roundR.}

\item{drop0}{Should trailing zeros be dropped?}

\item{groupvar}{Optional grouping variable for subgroups.}

\item{range}{Should min and max be included in output?}

\item{rangesep}{How should min/max be separated from mean+-sd?}

\item{.n}{Should n be included in output?}

\item{.german}{logical, should "." and "," be used as bigmark and decimal?}
}
\description{
Compute mean and sd and put together with the +- symbol.
}
