% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{meanse}
\alias{meanse}
\title{Compute mean and se and put together with the +- symbol.}
\usage{
meanse(x, mult = 1, roundDig = 2, drop0 = F)
}
\arguments{
\item{x}{Data for computation.}

\item{mult}{multiplier for se, default 1, can be set to 2 or 1.96 to create confidence intervals}

\item{roundDig}{Number of relevant digits for roundR.}

\item{drop0}{Should trailing zeros be dropped?}
}
\description{
Compute mean and se and put together with the +- symbol.
}
