% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{median_cl_boot}
\alias{median_cl_boot}
\title{Compute confidence interval of median by bootstrapping.}
\usage{
median_cl_boot(x, conf = 0.95, type = "basic")
}
\arguments{
\item{x}{Data for computation.}

\item{conf}{confidence interval with default 95\%.}

\item{type}{type for function boot.ci.}
}
\description{
Compute confidence interval of median by bootstrapping.
}
