% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{median_cl_boot_gg}
\alias{median_cl_boot_gg}
\title{Rename output from \link{median_cl_boot} for use in ggplot.}
\usage{
median_cl_boot_gg(x)
}
\arguments{
\item{x}{Data for computation.}
}
\value{
A tibble with one row and three columns: y, ymin, ymax.
}
\description{
\code{median_cl_boot_gg} computes lower and upper confidence limits for the
estimated median, based on bootstrapping, using default settings.
}
\examples{
# basic usage of median_cl_boot
median_cl_boot_gg(x = mtcars$wt)
}
