% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{medianse}
\alias{medianse}
\title{Compute standard error of median.}
\usage{
medianse(x)
}
\arguments{
\item{x}{Data for computation.}
}
\value{
numeric vector with SE Median.
}
\description{
\code{medianse} is based on \code{\link{mad}}/square root(n)
}
\examples{
# basic usage of medianse
medianse(x = mtcars$wt)
}
