% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{pdf_kable}
\alias{pdf_kable}
\title{Enhanced kable with latex}
\usage{
pdf_kable(.input, width1 = 6, twidth = 14, tposition = "left",
  innercaption = NULL, caption = "", foot = NULL, escape = T)
}
\arguments{
\item{.input}{table to print}

\item{twidth}{default 14}

\item{innercaption}{subheader}

\item{caption}{header}

\item{foot}{footnote}

\item{escape}{see kable}
}
\description{
\code{pdf_kable} formats tibbles/df's for markdown
}
