% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_ss.R
\name{pwr_ord}
\alias{pwr_ord}
\title{Computation of sample size for traits with ordinal distribution}
\usage{
pwr_ord(
  .lambda_ctr,
  .lambda_ctr_range = 1,
  .lambda_delta,
  .lambda_delta_range = 1,
  .lambda_step = 0.5,
  .nrepl = 10^3,
  .n_min = 4,
  .label_ctr = "Lambda control",
  .label_verum = "Lambda verum"
)
}
\arguments{
\item{.lambda_ctr}{expected count in control group.}

\item{.lambda_delta}{Expected difference on counts for treatment.}

\item{.lambda_delta_range}{Range for optimistic/pessimistic estimation for treatment. Within range, lambda is increased stepwise.}

\item{.lambda_step}{defines stepwise increase.}

\item{.nrepl}{Number of replications for simulation.}

\item{.n_min}{Smallest sample size to start with.}

\item{.label_ctr}{Label for plot.}

\item{.label_verum}{Label for plot.}

\item{.lamba_ctr_range}{Range for optimistic/pessimistic estimation for controls, resulting in three estimations.}
}
\description{
Computation of sample size for traits with ordinal distribution
}
