% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{SEM}
\alias{SEM}
\title{Standard Error of Mean.}
\usage{
SEM(x)
}
\arguments{
\item{x}{Data for computation.}
}
\value{
numeric vector with SEM.
}
\description{
\code{SEM} computes standard error of mean.
}
\examples{
SEM(x = mtcars$wt)
}
