% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{surprisal}
\alias{surprisal}
\title{Compute surprisal aka Shannon information from p-values}
\usage{
surprisal(p, precision = 1)
}
\arguments{
\item{p}{a vector of p-values}

\item{precision}{rounding level with default 1}
}
\value{
a character vector of s-values
}
\description{
\code{surprisal} takes p-values and returns s, a value representing the
number of consecutive heads on a fair coin, that would be as surprising
as the p-value
}
