% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{t_var_test}
\alias{t_var_test}
\title{Two independent sample t-test with decision for var.equal based on var.test}
\usage{
t_var_test(data, formula, cutoff = 0.05)
}
\arguments{
\item{data}{Tibble or data_frame.}

\item{formula}{Formula object with dependent and independent variable.}

\item{cutoff}{is significance threshold for equal variances}
}
\description{
Two independent sample t-test with decision for var.equal based on var.test
}
