% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{tab.search}
\alias{tab.search}
\title{Search within data.frame or tibble}
\usage{
tab.search(searchdata = rawdata, pattern, find.all = TRUE, names.only = FALSE)
}
\arguments{
\item{searchdata}{table to search in, predefined as rawdata}

\item{pattern}{regex, for exact matches add ^findme$}

\item{find.all}{return all row indices or only 1st per column, default = TRUE}

\item{names.only}{return only vector of colnames rather than list with names
and rows, default = FALSE}
}
\value{
A list with numeric vectors for each column giving row numbers
of matched elements
}
\description{
\code{tab.search} searches for pattern within a data-frame or tibble,
returning column(s) and row(s)
}
