% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{var_coeff}
\alias{var_coeff}
\title{Compute coefficient of variance.}
\usage{
var_coeff(x)
}
\arguments{
\item{x}{Data for computation.}
}
\value{
numeric vector with coefficient of variance.
}
\description{
\code{var_coeff computes relative variability as standard deviation/mean *100}
}
\examples{
var_coeff(x = mtcars$wt)
}
