% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alluvialplot}
\alias{alluvialplot}
\title{Create alluvialplot.}
\usage{
alluvialplot(
  .data,
  .x,
  .label = NULL,
  .fill = NULL,
  .title = NULL,
  .gridrow = NULL,
  .gridcol = NULL,
  .fillcolors = NULL
)
}
\arguments{
\item{.fill}{Variable in .data defining fill color.}
}
\description{
\code{alluvialplot} returns a ggplot object.
}
