% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{cat_desc_stats}
\alias{cat_desc_stats}
\title{Compute absolute and relative frequencies.}
\usage{
cat_desc_stats(quelle, trenner = "ARRR", return_level = TRUE,
  ndigit = 0, groupvar = NULL, singleline = FALSE, percent = TRUE)
}
\arguments{
\item{quelle}{Data for computation.}

\item{trenner}{delimiter between results per level. ARRR is my placeholder for later replacement with ^p (newline) in Word}

\item{return_level}{Should levels be reported?}

\item{ndigit}{Digits for rounding of relative frequencies.}

\item{percent}{logical, add % after relative frequencies?}
}
\description{
Compute absolute and relative frequencies.
}
