% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{cat_desc_stats}
\alias{cat_desc_stats}
\title{Compute absolute and relative frequencies.}
\usage{
cat_desc_stats(quelle, trenner = "ARRR", return_level = T,
  ndigit = 0)
}
\arguments{
\item{quelle}{Data for computation.}

\item{trenner}{delimiter between results per level.}

\item{return_level}{Should levels be reported?}

\item{ndigits}{Digits for rounding of %.}
}
\description{
Compute absolute and relative frequencies.
}
