% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{cn}
\alias{cn}
\title{Shortcut for colnames()}
\usage{
cn(data = rawdata)
}
\arguments{
\item{data}{Data structure to read column names from.}
}
\value{
Character vector with column names.
}
\description{
\code{cn} lists column names, by default for variable rawdata.
}
\examples{
cn(mtcars)
}
