% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare2numvars}
\alias{compare2numvars}
\title{Comparison for columns of numbers for 2 groups}
\usage{
compare2numvars(data, testvars, groupvar, gaussian, round_p = 3,
  round_desc = 2, range = F, pretext = F, mark = F)
}
\arguments{
\item{gaussian}{logical specifying normal or ordinal values.}
}
\description{
Comparison for columns of numbers for 2 groups
}
