% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare2numvars}
\alias{compare2numvars}
\title{Comparison for columns of numbers for 2 groups}
\usage{
compare2numvars(data, testvars, groupvar, gaussian, round_p = 3,
  round_desc = 2, range = F, pretext = F, mark = F, n = F,
  .n = F)
}
\arguments{
\item{data}{name of dataset (tibble/data.frame)to analyze.}

\item{testvars}{vector of column names.}

\item{groupvar}{name of grouping variable, has to translate to 2 groups.}

\item{gaussian}{logical specifying normal or ordinal values.}

\item{round_p}{level for rounding p-value}

\item{round_desc}{number of significant digits for rounding of descriptive stats}

\item{range}{include min/max?}

\item{pretext}{for function formatP}

\item{mark}{for function formatP}

\item{n}{create columns for n per group?}

\item{.n}{add n to descriptive statistics?}
}
\description{
Comparison for columns of numbers for 2 groups
}
