% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare_n_numvars}
\alias{compare_n_numvars}
\title{Comparison for columns of Gaussian measures for n groups}
\usage{
compare_n_numvars(
  .data = rawdata,
  testvars,
  groupvar,
  round_desc = 2,
  range = FALSE,
  rangesep = " ",
  pretext = FALSE,
  mark = FALSE,
  round_p = 3,
  .n = FALSE
)
}
\arguments{
\item{.data}{name of dataset (tibble/data.frame)to analyze, defaults to rawdata.}

\item{testvars}{vector of column names.}

\item{groupvar}{name of grouping variable.}

\item{round_desc}{number of significant digits for rounding of descriptive stats}

\item{range}{include min/max?}

\item{rangesep}{text between statistics and range or other elements}

\item{pretext}{for function formatP}

\item{mark}{for function formatP}

\item{round_p}{level for rounding p-value}

\item{.n}{add n to descriptive statistics?}
}
\description{
Comparison for columns of Gaussian measures for n groups
}
