% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare_n_qualvars}
\alias{compare_n_qualvars}
\title{Comparison for columns of factors for more than 2 groups with post-hoc}
\usage{
compare_n_qualvars(
  data,
  testvars,
  groupvar,
  round_p = 3,
  round_desc = 2,
  pretext = F,
  mark = F,
  singleline = F,
  spacer = "&nbsp;",
  linebreak = "\\n",
  prettynum = F
)
}
\arguments{
\item{data}{name of data set (tibble/data.frame) to analyze.}

\item{testvars}{vector of column names.}

\item{groupvar}{name of grouping variable, has to translate to 2 groups.}

\item{round_p}{level for rounding p-value.}

\item{round_desc}{number of significant digits for rounding of descriptive stats}

\item{pretext}{for function \link{formatP}}

\item{mark}{for function \link{formatP}}

\item{singleline}{Put all group levels in  a single line?}

\item{spacer}{Text element to indent levels.}

\item{linebreak}{place holder for newline.}

\item{prettynum}{Apply prettyNum to results?}
}
\description{
Comparison for columns of factors for more than 2 groups with post-hoc
}
