% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faketrial}
\alias{faketrial}
\title{Results from a simulated clinical trial with interaction effects.}
\format{
A tibble with 300 rows and 24 variables:
\describe{
\item{Sex}{Sex of animal, factor with levels 'female', 'male'}
\item{Agegroup}{Factor with levels 'young','middle','old'}
\item{Treatment}{Factor with levels 'sham', 'OP'}
\item{HR}{Heart rate}
\item{sysRR,diaRR}{Systolic and diastolic blood pressure}
\item{Med xxx}{Pseudo-medications, factors with levels 'y','n'}
\item{Biomarker x \link{units}}{Biomarkers with log-normal distribution}
}
}
\usage{
faketrial
}
\description{
A dataset containing physiological data, biomarkers, and categorical data.
}
\keyword{datasets}
