% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{flex2rmd}
\alias{flex2rmd}
\title{Transform flextable to rmd if non-interactive}
\usage{
flex2rmd(ft)
}
\arguments{
\item{ft}{a flextable}
}
\value{
either a markdown table from \code{flextable_to_rmd} or the printed flextable
}
\description{
\code{flex2rmd} takes a flextable and returns a markdown table if not in an interactive session, otherwise it prints the flextable. This is usefull e.g. in a loop.
}
