% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{formatP}
\alias{formatP}
\title{Re-format p-values}
\usage{
formatP(pIn, ndigits = 3, text = T, pretext = F, mark = F)
}
\arguments{
\item{pIn}{A numeric vector or matrix with p-values.}

\item{ndigits}{Number of digits (default=3).}

\item{text}{Should output be casted to character default=T)?}

\item{pretext}{Should = or < be added before p (default=F)?}

\item{mark}{Should significance level be added after p (default=F)?}
}
\description{
\code{formatP} simplifies p-values by rounding to the maximum of p or a
predefined level. Optionally < or = can be added, as well as
symbols according to significance level.
}
