% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{glmCI}
\alias{glmCI}
\title{Computation and formatting of CIs for glm}
\usage{
glmCI(model, min = 0.01, max = 100, cisep = "<U+22EF>", ndigit = 2)
}
\arguments{
\item{model}{Output from glm .}

\item{min}{Smallest OR to report.}

\item{max}{Largest OR to report.}

\item{cisep}{Separator between CI values.}

\item{ndigit}{rounding level.}
}
\description{
\code{glmCI} returns a list with coefficient,CIs,
and coef [CIs].
}
