% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ksnormal}
\alias{ksnormal}
\title{Kolmogorov-Smirnov-Test against Normal distribution}
\usage{
ksnormal(x, lillie = TRUE)
}
\arguments{
\item{x}{Vector of data to test.}

\item{lillie}{Logical, should the Lilliefors test be used? Defaults to TRUE}
}
\value{
p.value from \link{ks.test}.
}
\description{
\code{ksnormal} is a convenience function around \link{ks.test}, testing against
Normal distribution.
If less than 2 values are provided, NA is returned.
}
\examples{
# original ks.test:
ks.test(
  x = mtcars$wt, pnorm, mean = mean(mtcars$wt, na.rm = TRUE),
  sd = sd(mtcars$wt, na.rm = TRUE)
)
# wrapped version:
ksnormal(x = mtcars$wt, lillie = FALSE)
}
