% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{logrange_1}
\alias{logrange_1}
\title{vector for log-scales at 1}
\format{
An object of class \code{numeric} of length 41.
}
\usage{
logrange_1
}
\description{
vector for log-scales at 1
}
\keyword{datasets}
