% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{markSign}
\alias{markSign}
\title{Convert significance levels to symbols}
\usage{
markSign(SignIn, plabel = c("n.s.", "+", "*", "**", "***"))
}
\arguments{
\item{SignIn}{A single p-value.}

\item{plabel}{A translation table, predefined with the usual symbols}
}
\description{
\code{markSign} returns a single text element.
}
