% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{meansd}
\alias{meansd}
\title{Compute mean and sd and put together with the +- symbol.}
\usage{
meansd(
  x,
  roundDig = 2,
  drop0 = F,
  groupvar = NULL,
  range = F,
  rangesep = "ARRR",
  .n = F
)
}
\arguments{
\item{x}{Data for computation.}

\item{roundDig}{Number of relevant digits for roundR.}

\item{drop0}{Should trailing zeros be dropped?}

\item{groupvar}{Optional grouping variable for subgroups.}

\item{range}{Should min and max be included in output?}

\item{rangesep}{How should min/max be separated from mean+-sd?
Default ARRR is my shortcut for newline ^p later in Word.}

\item{.n}{Should n be included in output?}
}
\description{
Compute mean and sd and put together with the +- symbol.
}
\examples{
# basic usage of meansd
meansd(x=mtcars$wt)
# more elaborate usage of meansd
meansd(x=mtcars$wt, roundDig=3, groupvar=mtcars$gear,
.n=T, range=T,rangesep=" : ")
}
