% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{meansd}
\alias{meansd}
\title{Compute mean and sd and put together with the +- symbol.}
\usage{
meansd(x, roundDig = 2, drop0 = F, groupvar = NULL, range = F,
  rangesep = "ARRR")
}
\arguments{
\item{x}{Data for computation.}

\item{roundDig}{Number of relevant digigts for roundR.}

\item{drop0}{Should trailing zeros be dropped?}

\item{groupvar}{Optional grouping variable for subgroups.}

\item{range}{Should min and max be included in output?}

\item{rangesep}{How should min/max be separated from mean+-sd?
Default ARRR is my shortcut for newline ^p later in Word.}
}
\description{
Compute mean and sd and put together with the +- symbol.
}
