% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{pairwise_ordcat_test}
\alias{pairwise_ordcat_test}
\title{Calculate pairwise comparisons for ordinal categories between group levels
with corrections for multiple testing}
\usage{
pairwise_ordcat_test(
  x,
  group,
  adjmethod = "fdr",
  plevel = 0.05,
  symbols = letters[-1],
  ref = FALSE,
  method = "cmh"
)
}
\arguments{
\item{x}{dependent variable, containing the data}

\item{group}{independent variable, should be factor}

\item{adjmethod}{method for adjusting p values (see \link{p.adjust})}

\item{plevel}{threshold for significance}

\item{symbols}{predefined as b,c, d...;  provides footnotes to mark group
differences, e.g. b means different from group 2}

\item{ref}{is the 1st subgroup the reference (like in Dunnett test)}

\item{method}{which function should be used for individual testing?
Predefined to cmh from package coin
export}
}
\description{
Calculate pairwise comparisons for ordinal categories between group levels
with corrections for multiple testing
}
