% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{pairwise_wilcox_test}
\alias{pairwise_wilcox_test}
\title{Calculate pairwise comparisons between group levels with corrections for
multiple testing}
\usage{
pairwise_wilcox_test(
  dep_var,
  indep_var,
  strat_var = NA,
  adjmethod = "fdr",
  distr = "exact",
  plevel = 0.05,
  symbols = letters[-1],
  sep = ""
)
}
\arguments{
\item{dep_var}{dependent variable, containing the data}

\item{indep_var}{independent variable, should be factor}

\item{strat_var}{optiona factor for stratification}

\item{adjmethod}{method for adjusting p values (see \link{p.adjust})}

\item{distr}{The conditional null distribution of the test statistic is used to obtain p-values and an asymptotic approximation of the exact distribution is used by default (distribution = "asymptotic"). Alternatively, the distribution can be approximated via Monte Carlo resampling or computed exactly for univariate two-sample problems by setting distribution to "approximate" or "exact" respectively.}

\item{plevel}{threshold for significance}

\item{symbols}{predefined as b,c, d...;  provides footnotes to mark group
differences, e.g. b means different from group 2}

\item{sep}{text between statistics and range or other elements}
}
\description{
Calculate pairwise comparisons between group levels with corrections for
multiple testing
}
