% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{pdf_kable}
\alias{pdf_kable}
\title{Enhanced \link[knitr:kable]{knitr::kable} with latex}
\usage{
pdf_kable(
  .input,
  width1 = 6,
  twidth = 14,
  tposition = "left",
  innercaption = NULL,
  caption = "",
  foot = NULL,
  escape = TRUE
)
}
\arguments{
\item{.input}{table to print}

\item{width1}{Width of 1st column, default 6.}

\item{twidth}{Default 14}

\item{tposition}{Default left}

\item{innercaption}{subheader}

\item{caption}{header}

\item{foot}{footnote}

\item{escape}{see \link[knitr:kable]{knitr::kable}}
}
\value{
A character vector of the table source code.
}
\description{
\code{pdf_kable} formats tibbles/df's for markdown
}
