% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotech.R
\name{plot_LB}
\alias{plot_LB}
\title{Lineweaver-Burk diagram}
\usage{
plot_LB(
  substrate,
  velocity,
  title = "Lineweaver-Burk-Plot",
  xlab = "1/substrate",
  ylab = "1/velocity"
)
}
\arguments{
\item{substrate}{substrate concentration}

\item{velocity}{enzyme velocity}

\item{title}{title of the plot}

\item{xlab}{label of the abscissa}

\item{ylab}{label of the ordinate}
}
\description{
\code{plot_LB} plots a Lineweaver-Burk diagram and computes the linear model
}
\examples{
subst <- c(2.00, 1.00, 0.50, 0.25)
velo <- c(0.2253, 0.1795, 0.1380, 0.1000)
plot_LB(substrate = subst,velocity = velo)

}
