% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotech.R
\name{plot_MM}
\alias{plot_MM}
\title{Michaelis-Menten enzyme kinetics model and plot}
\usage{
plot_MM(
  substrate,
  velocity,
  title = "Michaelis-Menten",
  xlab = "substrate",
  ylab = "velocity"
)
}
\arguments{
\item{substrate}{substrate concentration}

\item{velocity}{reaction velocity}

\item{title}{title of the plot}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}
}
\value{
a list with elements "MMfit" and "MMplot"
}
\description{
\code{plot_MM} creates a Michaelis-Menten type Enzyme kinetics plot and returns model as well
}
\examples{
subst <- c(2.00, 1.00, 0.50, 0.25)
velo <- c(0.2253, 0.1795, 0.1380, 0.1000)
plot_MM(substrate = subst,velocity = velo)

}
