% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{print_kable}
\alias{print_kable}
\title{Enhanced kable with definable number of rows/columns for splitting}
\usage{
print_kable(t, nrows = 30, caption = "", ncols = 100, ...)
}
\arguments{
\item{t}{table to print.}

\item{nrows}{number of rows (30) before splitting.}

\item{caption}{header.}

\item{ncols}{number of columns (100) before splitting.}

\item{...}{Further arguments passed to \link{kable}.}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
package flextable is a more powerful alternative

\code{print_kable} formats and prints tibbles/df's in markdown with splitting
into sub-tables with repeated caption and header.
}
\examples{
\dontrun{
print_kable(mtcars, caption = "test")
}
}
