% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{roundR}
\alias{roundR}
\title{Automatic rounding to a reasonable length, based on largest number}
\usage{
roundR(
  roundin,
  smooth = F,
  level = 2,
  textout = T,
  drop0 = F,
  .german = F,
  .bigmark = F
)
}
\arguments{
\item{roundin}{A vector or matrix of numbers.}

\item{smooth}{A logical specifying if you want rounding before the dot
(e.g. 12345 to 12300).}

\item{level}{A number specifying number of relevant digits to keep.}

\item{textout}{A logical if output is converted to text.}

\item{drop0}{A logical if trailing zeros should be dropped}

\item{.german}{A logical if german numbers should be reported}

\item{.bigmark}{A logical if big.mark is to be shown, mark itself
depends on parameter .german}
}
\value{
vector of type character or numeric, depending on parameter textout
}
\description{
\code{roundR} returns a matrix of rounded numbers.
}
