% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{se_median}
\alias{se_median}
\title{Compute standard error of median.}
\usage{
se_median(x)
}
\arguments{
\item{x}{Data for computation.}
}
\value{
numeric vector with SE Median.
}
\description{
\code{median_cl_boot} is based on \code{\link{mad}}/square root(n)
}
\examples{
# basic usage of meanse
se_median(x = mtcars$wt)
}
