% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefunctions.R
\name{subchunkify}
\alias{subchunkify}
\title{Change figure size within chunk, taken from http://michaeljw.com/blog/post/subchunkify/}
\usage{
subchunkify(g, fig_height = 7, fig_width = 5)
}
\arguments{
\item{g}{plot object}
}
\description{
Change figure size within chunk, taken from http://michaeljw.com/blog/post/subchunkify/
}
