% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedtools-package.R
\docType{package}
\name{wrappedtools-package}
\alias{wrappedtools}
\alias{wrappedtools-package}
\title{wrappedtools: Useful Wrappers Around Commonly Used Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The main functionalities of 'wrappedtools' are: adding backticks to variable names; rounding to desired precision with special case for p-values; selecting columns based on pattern and storing their position, name, and backticked name; computing and formatting of descriptive statistics (e.g. mean±SD), comparing groups and creating publication-ready tables with descriptive statistics and p-values; creating specialized plots for correlation matrices. Functions were mainly written for my own daily work or teaching, but may be of use to others as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abusjahn/wrappedtools}
  \item Report bugs at \url{https://github.com/abusjahn/wrappedtools/issues}
}

}
\author{
\strong{Maintainer}: Andreas Busjahn \email{andreas@busjahn.net} (\href{https://orcid.org/0000-0001-9650-6919}{ORCID})

Authors:
\itemize{
  \item Bilal Asser \email{billyasser@hotmail.co.uk}
  \item Franziska Eidloth \email{franziska.eidloth@gmail.com}
}

}
\keyword{internal}
